requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
Flask-Cors==5.0.0
flask-marshmallow==1.3.0
Flask-PyMongo==2.3.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
marshmallow==3.25.1
packaging==24.2
pluggy==1.5.0
pymongo==3.13.0
pymongo-stubs==0.2.0
pytest==8.3.4
pytest-flask==1.3.0
pytest-mock==3.14.0
python-dotenv==1.0.1
Werkzeug==3.1.3
